feat: add experimental version of Lit based details
