+++ packages/details/test/details-lit.test.js
@@ -1,0 +1,2 @@
+import '../src/vaadin-lit-details.js';
+import './details.common.js';
