+++ packages/details/test/details-polymer.test.js
@@ -1,0 +1,2 @@
+import '../src/vaadin-details.js';
+import './details.common.js';
